+++ tap_ga4/__init__.py
@@ -1,3 +1,4 @@
+import json
 import singer
 from singer import utils
 from singer.catalog import Catalog
@@ -14,22 +15,30 @@ REQUIRED_CONFIG_KEYS = [
     "refresh_token",
     "property_id",
     "account_id",
-    "report_definitions",
 ]
 
+def maybe_parse_report_definitions(config):
+    """Converts report_definitions into a list if it is a JSON-encoded string."""
+    if isinstance(config.get("report_definitions", []), str):
+        try:
+            config.update(report_definitions = json.loads(config["report_definitions"]))
+        except json.JSONDecodeError as e:
+            raise ValueError(f"Error parsing report_definitions string: {e}") from e
 
 def main_impl():
     args = utils.parse_args(REQUIRED_CONFIG_KEYS)
-    config = args.config
     catalog = args.catalog or Catalog([])
     state = {}
+
+    config = args.config
+    maybe_parse_report_definitions(config)
 
     client = Client(config)
 
     if args.state:
         state.update(args.state)
     if args.discover:
-        discover(client, config["report_definitions"], config["property_id"])
+        discover(client, config.get("report_definitions", []), config["property_id"])
         LOGGER.info("Discovery complete")
     elif args.catalog:
         sync(client, config, catalog, state)